MNT: move piecewise functions to separate file (#746)
* MNT: move piecewise functions to separate file
closes #667

* improved import for linting

* MNT: applying code formaters

* ENH: simplifying and optimizing the function, implementing tests.

* MNT: update changelog and apply changes suggested in review

---------
