+++ rocketpy/mathutils/__init__.py
@@ -1,7 +1,3 @@
-from .function import (
-    Function,
-    PiecewiseFunction,
-    funcify_method,
-    reset_funcified_methods,
-)
+from .function import Function, funcify_method, reset_funcified_methods
+from .piecewise_function import PiecewiseFunction
 from .vector_matrix import Matrix, Vector
