+++ rocketpy/motors/tank_geometry.py
@@ -2,7 +2,8 @@ from functools import cached_property
 
 import numpy as np
 
-from ..mathutils.function import Function, PiecewiseFunction, funcify_method
+from ..mathutils.function import Function, funcify_method
+from ..mathutils.piecewise_function import PiecewiseFunction
 from ..plots.tank_geometry_plots import _TankGeometryPlots
 from ..prints.tank_geometry_prints import _TankGeometryPrints
